# booking-service/src/swagger.json
{
  "openapi": "3.0.0",
  "info": { "title": "Booking Service API", "version": "1.0.0" },
  "servers": [{ "url": "http://localhost:3002" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/resources": {
      "post": {
        "summary": "Create a resource (Admin only)",
        "tags": ["Resources"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Conference Room A" },
                  "description": { "type": "string", "example": "First floor meeting room" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Resource created" }, "409": { "description": "Already exists" } }
      },
      "get": {
        "summary": "List all active resources",
        "tags": ["Resources"],
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "List of resources" } }
      }
    },
    "/bookings": {
      "post": {
        "summary": "Create a booking",
        "tags": ["Bookings"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "resource": { "type": "string", "example": "uuid-of-resource" },
                  "startTime": { "type": "string", "example": "2025-09-01T10:00:00Z" },
                  "endTime": { "type": "string", "example": "2025-09-01T11:00:00Z" }
                },
                "required": ["resource", "startTime", "endTime"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Booking created" } }
      },
      "get": {
        "summary": "List bookings (user sees own, admin sees all)",
        "tags": ["Bookings"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "example": "BOOKED" } },
          { "name": "resourceId", "in": "query", "schema": { "type": "string", "example": "uuid-of-resource" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "example": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "example": 10 } }
        ],
        "responses": { "200": { "description": "List of bookings" } }
      }
    },
    "/bookings/{id}": {
      "get": {
        "summary": "Get booking by ID (own or any for admin)",
        "tags": ["Bookings"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "uuid-here" } }
        ],
        "responses": { "200": { "description": "Booking found" } }
      },
      "patch": {
        "summary": "Update booking (own or any for admin)",
        "tags": ["Bookings"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "uuid-here" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "resourceId": { "type": "string", "example": "uuid-of-new-resource" },
                  "startTime": { "type": "string", "example": "2025-09-01T12:00:00Z" },
                  "endTime": { "type": "string", "example": "2025-09-01T13:00:00Z" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Booking updated" } }
      },
      "delete": {
        "summary": "Cancel booking (own or any for admin)",
        "tags": ["Bookings"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "example": "uuid-here" } }
        ],
        "responses": { "200": { "description": "Booking cancelled" } }
      }
    },
    "/resources/{id}": {
      "patch": {
        "summary": "Update a resource (Admin only)",
        "tags": ["Resources"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "uuid-of-resource" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Updated Room A" },
                  "description": { "type": "string", "example": "Renovated and bigger" },
                  "isActive": { "type": "boolean", "example": false }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Resource updated" },
          "404": { "description": "Not found" }
        }
      }
    }
  }
}